requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
distro==1.9.0
filelock==3.20.3
fsspec==2026.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.187.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-genai==1.57.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.3
langchain-core==1.2.7
langchain-openai==1.1.7
langchain-postgres==0.0.16
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.2
langsmith==0.6.2
markupsafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.1
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai==2.15.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
pdf2image==1.17.0
pgvector==0.3.6
pillow==12.1.0
proto-plus==1.27.0
protobuf==5.29.5
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg-pool==3.3.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.12.5
pydantic-core==2.41.5
pyparsing==3.3.1
pypdf2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyyaml==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
sqlalchemy==2.0.45
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
triton==3.5.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uuid-utils==0.13.0
websockets==15.0.1
xxhash==3.6.0
zstandard==0.25.0
